fix: various bugs
