+++ frontend/src/pages/rag/[id]/add-documents.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { useRouter } from 'next/router'
 import AddDocumentsWizard from '@/components/rag/AddDocumentsWizard'
+import Header from '@/components/Header'
 
 const AddDocumentsPage: React.FC = () => {
   const router = useRouter()
@@ -10,7 +11,12 @@ const AddDocumentsPage: React.FC = () => {
     return null
   }
 
-  return <AddDocumentsWizard knowledgeBaseId={id} />
+  return (
+    <div className="App relative">
+      <Header activePage="rag" />
+      <AddDocumentsWizard knowledgeBaseId={id} />
+    </div>
+  )
 }
 
 export default AddDocumentsPage
