+++ frontend/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "target": "es2016",
-        "module": "commonjs",
+        "module": "esnext",
         "baseUrl": ".",
         "plugins": [
             {
@@ -9,24 +9,39 @@
             }
         ],
         "paths": {
-            "@/*": ["./src/*"]
+            "@/*": [
+                "./src/*"
+            ]
         },
         "esModuleInterop": true,
         "forceConsistentCasingInFileNames": true,
         "strict": false,
         "skipLibCheck": true,
-        "lib": ["dom", "dom.iterable", "esnext"],
+        "lib": [
+            "dom",
+            "dom.iterable",
+            "esnext"
+        ],
         "allowJs": true,
         "checkJs": true,
         "noEmit": true,
         "jsx": "preserve",
-        "moduleResolution": "node",
+        "moduleResolution": "bundler",
         "resolveJsonModule": true,
         "isolatedModules": true,
         "incremental": true,
         "noErrorTruncation": true,
         "preserveWatchOutput": true
     },
-    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/**/*", "postcss.config.js", "tailwind.config.js"],
-    "exclude": ["node_modules"]
+    "include": [
+        "next-env.d.ts",
+        "**/*.ts",
+        "**/*.tsx",
+        "src/**/*",
+        "postcss.config.js",
+        "tailwind.config.js"
+    ],
+    "exclude": [
+        "node_modules"
+    ]
 }
